Add an extra layer to handle direct writes to the end of .bashrc

+++ install.sh
@@ -2,10 +2,17 @@
 
 set -e
 
+ensure_line_exists() {
+  grep -q -F "$1" "$2" \
+    || echo "$1" >> "$2"
+}
+
 install_bash() {
-  [ -L $HOME/.bashrc ] && (echo "There is already an symlink at .bashrc"; return )
+  [ -L $HOME/.bashrc ] && { echo "There is already an symlink at .bashrc"; exit 0; }
   [ -f $HOME/.bashrc ] && mv $HOME/.bashrc $HOME/.bashrc.backup
-  ln -s $(pwd)/bash/entry.in.sh $HOME/.bashrc
+  touch $(pwd)/bash/bashrc
+  ensure_line_exists ". $(pwd)/bash/entry.in.sh" $(pwd)/bash/bashrc
+  ln -s $(pwd)/bash/bashrc $HOME/.bashrc
 }
 
 main() {
